Update Blackjack 2
The functionality of the game did not change. This update limited lines to 79 characters and comment lines to 72 characters to follow the PEP 8 guidelines. I also tried my best to follow other PEP 8 guidelines, but I'm sure I'm missing some things. Nevertheless, this should prove to be more readable to everyone.